Updated input model
